# Switch to github.com/oasisprotocol/curve25519-voi