--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,6 @@ module github.com/AlexanderYastrebov/wireguard-vanity-key
 
 go 1.23.2
 
-require filippo.io/edwards25519 v1.1.1-0.20250126203914-62e99b0d908f
+require github.com/oasisprotocol/curve25519-voi v0.0.0-20230904125328-1f23a7beb09a
+
+require golang.org/x/sys v0.0.0-20220325203850-36772127a21f // indirect