--- a/main_test.go
+++ b/main_test.go
@@ -8,7 +8,7 @@ import (
 	"sync/atomic"
 	"testing"
 
-	"filippo.io/edwards25519"
+	"github.com/oasisprotocol/curve25519-voi/curve"
 )
 
 func BenchmarkGenerateKey(b *testing.B) {
@@ -23,7 +23,7 @@ func BenchmarkFindPoint(b *testing.B) {
 
 	i := b.N
 
-	findPoint(context.Background(), p0, randUint64(), func(p *edwards25519.Point) bool {
+	findPoint(context.Background(), p0, randUint64(), func(p *curve.EdwardsPoint) bool {
 		match := hasBase64Prefix(p, []byte("GoodLuckWithThisPrefix"))
 		i--
 		return i == 0 || match
@@ -36,7 +36,7 @@ func BenchmarkFindPointParallel(b *testing.B) {
 	var i atomic.Int64
 	i.Store(int64(b.N))
 
-	findPointParallel(context.Background(), runtime.NumCPU(), p0, func(p *edwards25519.Point) bool {
+	findPointParallel(context.Background(), runtime.NumCPU(), p0, func(p *curve.EdwardsPoint) bool {
 		match := hasBase64Prefix(p, []byte("GoodLuckWithThisPrefix"))
 		return i.Add(-1) <= 0 || match
 	})