Use https://github.com/oasisprotocol/curve25519-voi/ which provides faster BytesMontgomery conversion due to use of Pow2k in the Element.Invert.

```
goos: linux
goarch: amd64
pkg: github.com/AlexanderYastrebov/wireguard-vanity-key
cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
                    │    main     │                 HEAD                 │
                    │   sec/op    │    sec/op     vs base                │
FindPointParallel-8   889.2n ± 6%   798.9n ± 13%  -10.15% (p=0.019 n=10)

                    │    main    │              HEAD              │
                    │    B/op    │    B/op     vs base            │
FindPointParallel-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                    │    main    │              HEAD              │
                    │ allocs/op  │ allocs/op   vs base            │
FindPointParallel-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```

See also https://github.com/FiloSottile/edwards25519/issues/23